[module7-task1]: Extract comments.router
